pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ranksentinel"
version = "0.1.0"
description = "SEO regression monitoring with weekly reports and low-noise critical alerts."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "google-analytics-data>=0.18.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
# This repo uses a "src/" layout; ensure tests can import "ranksentinel" without
# requiring an editable install.
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
